LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/quirks-auto-block-size-with-percentage-item-expected.txt

There should be a green square to the left of a blue square, and no red.



FAIL #container 1 assert_equals: 
<div id="container" style="width:200px; height:456px;">
  <div style="display:flex; background:blue;" data-expected-height="100">
    <img style="width:100px; height: 50%;" src="support/1x1-green.png" data-expected-height="100">
    <div style="width: 50px; height: 100%; background: red;" data-expected-height="0"></div>
  </div>
</div>
height expected 100 but got 456
